tools/translation_agent/requirements.txt
openpyxl
gspread
google-auth
gitpython
gspread-formatting
python-dotenv>=1.0.0
openai==2.8.0
openai-agents==0.5.0
pyyaml>=6.0